Update to Latest on local store.

+++ cleanNightly.sh
@@ -2,6 +2,10 @@
 #set -x
 
 version=$1
+BaseDownloadNFSDir="/home/data/httpd/download.eclipse.org/rt/eclipselink"
+DownloadDir=${BaseDownloadNFSDir}/nightly/${version}
+buildir=/shared/rt/eclipselink
+
 
 unset usage
 usage() {
@@ -10,18 +14,37 @@ usage() {
     echo "  version  Name of version for which old builds will be cleaned."
 }
 
+unset validateVersion
+validateVersion() {
+    runVersion=$1
+
+    cd ${BaseDownloadNFSDir}/nightly
+    for validVersion in `ls -dr [0-9]*` ; do
+        if [ "${validVersion}" = "${runVersion}" ] ; then
+            echo "Version found. ${runVersion} valid..."
+            valid=true
+        fi
+    done
+    if [ ! "$valid" = "true" ] ; then
+       echo "Version not present in nightly. '${runVersion}' isn't a valid or active version..."
+       exit
+    fi
+}
+
+##############################################
+#
+# MAIN
+#
 if [ \( -z "$version" \) -o \( "$version" = "" \) ] ; then
     usage
     exit
 else
+   validateVersion ${version}
    echo "Purging old builds of '${version}'..."
 fi
 
-BaseDownloadNFSDir="/home/data/httpd/download.eclipse.org/rt/eclipselink"
-buildir=/shared/rt/eclipselink
-
 cd ${buildir}
-if [ "input" == "release" ]
+if [ "input" = "release" ]
 then
     # When releasing clear all nightly builds
     num_builds=0
@@ -36,18 +59,24 @@ fi
 
 ### Download Site ###
 #      leave only the last 10 build dirs for the version on the download server
-index=0
-removed=0
-cd ${BaseDownloadNFSDir}/nightly/${version}
-for contentdir in `ls -dr [0-9]*` ; do
-    index=`expr $index + 1`
-    if [ $index -gt $num_builds ] ; then
-        echo "Removing ${contentdir}..."
-        rm -r $contentdir
-        removed=`expr $removed + 1`
-    fi
-done
-echo "Removed $removed direcories from ${BaseDownloadNFSDir}/nightly/${version}."
+if [ -d ${DownloadDir} ] ; then
+    index=0
+    removed=0
+    cd ${DownloadDir}
+    for contentdir in `ls -dr [0-9]*` ; do
+        index=`expr $index + 1`
+        if [ $index -gt $num_builds ] ; then
+            echo "Removing ${contentdir}..."
+            rm -r $contentdir
+            removed=`expr $removed + 1`
+        fi
+    done
+    echo "Removed $removed directories from ${BaseDownloadNFSDir}/nightly/${version}."
+else
+    echo "No '${BaseDownloadNFSDir}/nightly/${version}' dir found!
+    echo "    Assuming invalid version number, and aborting clean.
+    exit
+fi
 
 ### P2 Site ###
 #      leave only the last "num_p2_builds" builds for the version in the nightly P2 repos
@@ -64,6 +93,9 @@ for contentdir in `ls -dr ${version}*` ; do
 done
 echo "Removed $removed direcories from ${BaseDownloadNFSDir}/nightly-updates."
 
+echo "TODO: Need to verify the "correct" way to clean the maven repo of old SNAPSHOTs"
+echo "Hit the fullstop!"
+exit
 ### Maven Site ###
 #      leave only last 5 days worth of files in the maven repository
 cd ${BaseDownloadNFSDir}/maven.repo/org/eclipse/persistence
